edge runtime error fixed
