+++ app/(user)/(routes)/(auth)/sign-in/page.tsx
@@ -54,7 +54,7 @@ const SignIn = () => {
     <div className="flex items-center justify-center w-full h-[491px]   relative top-[60px] ">
       {/* Form Elements */}
       <div className="w-[385px] bg-[#1B1B1B] max-h-[423px] pl-8 pr-[103px] py-[20px] ">
-        <h3 className="text-white">Room'a Ücretsiz Kayıt Ol</h3>
+        <h3 className="text-white">Room&apos;a Ücretsiz Kayıt Ol</h3>
         {/* Form For email/password/username */}
         <form onSubmit={handleSubmit}>
           {/* email input  */}
@@ -81,7 +81,7 @@ const SignIn = () => {
                 type="email"
                 name="email"
                 id="email"
-                placeholder="Kullanıcı adınızı giriniz"
+                placeholder="Emailinizi giriniz"
                 className="px-[10px] py-2 rounded-[5px] placeholder:text-paragraphGray text-white bg-[transparent!important] focus:outline-none"
                 autoFocus
               />
@@ -122,7 +122,7 @@ const SignIn = () => {
                 type="password"
                 name="password"
                 id="password"
-                placeholder="Kullanıcı adınızı giriniz"
+                placeholder="Şifrenizi giriniz"
                 className="px-[10px] py-2 rounded-[5px] placeholder:text-paragraphGray text-white bg-[transparent!important] focus:outline-none"
                 autoFocus
               />
