+++ app/api/auth/sign-in/route.ts
@@ -62,3 +62,5 @@ export async function POST(request: Request) {
     { status: 400 }
   );
 }
+
+// export const runtime = "nodejs";
