+++ app/api/auth/sign-up/route.ts
@@ -173,3 +173,5 @@ export async function POST(request: Request) {
       break;
   }
 }
+
+// export const runtime = "nodejs";
